requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.2.4
django-cors-headers==3.1.0
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djoser==2.0.2
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
